+++ app/__init__.py
@@ -1,9 +1,22 @@
 from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
+
+db = SQLAlchemy()
 
 def create_app():
     app = Flask(__name__)
 
-    # Registrar blueprints
+    # Get the database URL from environment variables
+    app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv('DATABASE_URL')
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+
+    db.init_app(app)
+
+    # Import and register blueprints
     from app.api.routes import api_bp
     app.register_blueprint(api_bp, url_prefix='/api')
