+++ app/api/routes.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request, jsonify
-from app.models.models import analyze_sentiment
+from app.utils.FeedbackAnalysisPipeline import analyze_sentiment
+from app.models.models import Feedback, db
 
-# Criar um blueprint para as rotas da API
 api_bp = Blueprint('api', __name__)
 
 @api_bp.route('/feedbacks', methods=['POST'])
@@ -10,19 +10,25 @@ def analyze_feedback():
     feedback_id = data.get('id')
     feedback_text = data.get('feedback')
 
-    # Realizar análise de sentimento
+    # Perform sentiment analysis
     sentiment, requested_features = analyze_sentiment(feedback_text)
 
-    # Retornar a resposta com o formato desejado
+    # Save feedback to database
+    feedback = Feedback(
+        feedback_id=feedback_id,
+        feedback_text=feedback_text,
+        sentiment=sentiment['sentiment'],
+        code=requested_features.get('code'),
+        reason=requested_features.get('reason')
+    )
+    db.session.add(feedback)
+    db.session.commit()
+
+    # Return the response in the desired format
     response_data = {
         "id": feedback_id,
+        "sentiment": sentiment['sentiment'],
         "requested_features": requested_features
     }
-    response_data.update(sentiment)
-
-    print(response_data)
 
     return jsonify(response_data)
-
-if __name__ == "__main__":
-    api_bp.run(debug=True)
