+++ requirements.txt
@@ -4,3 +4,6 @@ python-dotenv==1.0.1
 google-generativeai==0.5.4
 langchain-google-genai==1.0.5
 pytest==8.2.1
+flask-mysql==1.5.2
+flask_sqlalchemy==3.1.1
+mysqlclient==2.2.4
